+++ CHANGELOG.md
@@ -6,6 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - TODO
+### Changed
+- Extracted email verification as its own recipe
+- ThirdParty queries
+
 ## [1.6.0] - 2021-01-14
 ### Changed
 - Used rowmapper interface